CMake: Copied and adjusted the Aseprite CMakeLists.txt files
By only including the relevant bits, we can configure the project much
faster.

Also mentioned the USE_SHARED_LIBPNG option. libpng is a dependency of
FreeType (normally the shared libpng would be used anyway, when
USE_SHARED_FREETYPE was used).

This is a first step towards being able to package a source release
using just the relevant parts of the repository (issue #12).